% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsegl-methods.R
\name{coef.sparsegl}
\alias{coef.sparsegl}
\title{Extract model coefficients from a `sparsegl` object.}
\usage{
\method{coef}{sparsegl}(object, s = NULL, ...)
}
\arguments{
\item{object}{Fitted [sparsegl()] object.}

\item{s}{Value(s) of the penalty parameter `lambda` at which
coefficients are required. Default is the entire sequence.}

\item{...}{Not used.}
}
\value{
The coefficients at the requested values for `lambda`.
}
\description{
Computes the coefficients at the requested value(s) for `lambda` from a
[sparsegl()] object.
}
\details{
`s` is the new vector of `lambda` values at which predictions are requested.
If `s` is not in the lambda sequence used for fitting the model, the `coef`
function will use linear interpolation to make predictions. The new values
are interpolated using a fraction of coefficients from both left and right
`lambda` indices.
}
\examples{
n <- 100
p <- 20
X <- matrix(rnorm(n * p), nrow = n)
eps <- rnorm(n)
beta_star <- c(rep(5, 5), c(5, -5, 2, 0, 0), rep(-5, 5), rep(0, (p - 15)))
y <- X \%*\% beta_star + eps
groups <- rep(1:(p / 5), each = 5)
fit1 <- sparsegl(X, y, group = groups)
coef(fit1, s = c(0.02, 0.03))
}
\seealso{
[sparsegl()] and [predict.sparsegl()].
}
