% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.sparsegl}
\alias{cv.sparsegl}
\title{Cross-validation for a `sparsegl` object.}
\usage{
cv.sparsegl(
  x,
  y,
  group = NULL,
  family = c("gaussian", "binomial"),
  lambda = NULL,
  pred.loss = c("default", "mse", "deviance", "mae", "misclass", "ROC"),
  nfolds = 10,
  foldid = NULL,
  weights = NULL,
  offset = NULL,
  cn,
  drgix,
  drgiy,
  cn_s,
  cn_e,
  partition = c("general", "subgroup"),
  subgroupnumber,
  ...
)
}
\arguments{
\item{pred.loss}{Loss to use for cross-validation error. Valid options are:
* `"default"` the same as deviance (mse for regression and deviance otherwise)
* `"mse"` mean square error
* `"deviance"` the default (mse for Gaussian regression, and negative
  log-likelihood otherwise)
* `"mae"` mean absolute error, can apply to any family
* `"misclass"` for classification only, misclassification error.}

\item{nfolds}{Number of folds - default is 10. Although `nfolds` can be
as large as the sample size (leave-one-out CV), it is not recommended for
large datasets. Smallest value allowable is `nfolds = 3`.}

\item{foldid}{An optional vector of values between 1 and `nfolds`
identifying which fold each observation is in. If supplied, `nfolds` can
be missing.}

\item{...}{Additional arguments to [sparsegl()].}
}
\value{
An object of class [cv.sparsegl()] is returned, which is a
  list with the components describing the cross-validation error.
  \item{lambda}{The values of \code{lambda} used in the fits.}
  \item{cvm}{The mean cross-validated error - a vector of
    length \code{length(lambda)}.}
  \item{cvsd}{Estimate of standard error of \code{cvm}.}
  \item{cvupper}{Upper curve = \code{cvm + cvsd}.}
  \item{cvlower}{Lower curve = \code{cvm - cvsd}.}
  \item{name}{A text string indicating type of measure (for plotting
    purposes).}
  \item{nnzero}{The number of non-zero coefficients for each \code{lambda}}
  \item{active_grps}{The number of active groups for each \code{lambda}}
  \item{sparsegl.fit}{A fitted [sparsegl()] object for the full data.}
  \item{lambda.min}{The optimal value of \code{lambda} that gives
    minimum cross validation error \code{cvm}.}
  \item{lambda.1se}{The largest value of \code{lambda} such that error
    is within 1 standard error of the minimum.}
  \item{call}{The function call.}
}
\description{
Performs k-fold cross-validation for [sparsegl()].
This function is largely similar [glmnet::cv.glmnet()].
}
\details{
The function runs [sparsegl()] `nfolds + 1` times; the first to
get the `lambda` sequence, and then the remainder to compute the fit
with each of the folds omitted. The average error and standard error
over the folds are computed.
}
\examples{
n <- 100
p <- 20
X <- matrix(rnorm(n * p), nrow = n)
eps <- rnorm(n)
beta_star <- c(rep(5, 5), c(5, -5, 2, 0, 0), rep(-5, 5), rep(0, (p - 15)))
y <- X \%*\% beta_star + eps
groups <- rep(1:(p / 5), each = 5)
cv_fit <- cv.sparsegl(X, y, groups)

}
\seealso{
[sparsegl()], as well as [`plot()`][plot.cv.sparsegl()],
  [`predict()`][predict.cv.sparsegl()], and [`coef()`][coef.cv.sparsegl()]
  methods for `"cv.sparsegl"` objects.
}
