% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{hierNest-package}
\alias{hierNest-package}
\title{hierNest: Penalized Regression with Hierarchical Nested Parameterization Structure}
\description{
Efficient implementation of penalized regression with hierarchical nested parametrization for grouped data. The package provides penalized regression methods that decompose subgroup-specific covariate effects into shared global effects, Major subgroup–specific effects, and Minor subgroup–specific effects, enabling structured borrowing of information across related clinical subgroups. Both lasso and hierarchical overlapping group lasso penalties are supported to encourage sparsity while respecting the nested subgroup structure. Efficient computation is achieved through a modified design matrix representation and a custom majorization–minimization algorithm for overlapping group penalties.
}
\author{
\strong{Maintainer}: Ziren Jiang \email{jian0746@umn.edu}

Authors:
\itemize{
  \item Jared Huling \email{huling@umn.edu}
  \item Jue Hou \email{hou00123@umn.edu}
  \item Lingfeng Huo
}

Other contributors:
\itemize{
  \item Daniel J. McDonald [contributor]
  \item Xiaoxuan Liang [contributor]
  \item Anibal Solón Heinsfeld [contributor]
  \item Aaron Cohen [contributor]
  \item Yi Yang [contributor]
  \item Hui Zou [contributor]
  \item Jerome Friedman [contributor]
  \item Trevor Hastie [contributor]
  \item Rob Tibshirani [contributor]
  \item Balasubramanian Narasimhan [contributor]
  \item Kenneth Tay [contributor]
  \item Noah Simon [contributor]
  \item Junyang Qian [contributor]
  \item James Yang [contributor]
}

}
\keyword{internal}
