% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{hierNest-package}
\alias{hierNest-package}
\title{hierNest: Penalized regression with hierarchical nested parametrization sturcture}
\description{
Efficient implementation of penalized regression with hierarchical nested parametrization for grouped data.
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Ziren Jiang [contributor]
  \item Jared Huling [contributor]
  \item Jue Hou [contributor]
  \item Lingfeng Huo [contributor]
}

}
